[MOV] sale_purchase: split sale_order.py
