+++ addons/sale_purchase/models/__init__.py
@@ -3,4 +3,5 @@
 
 from . import product_template
 from . import sale_order
+from . import sale_order_line
 from . import purchase_order
